requirements.txt
pyside6
pyqt6-tools
pyqt6rc
